requirements.txt
# MCP Server Dependencies
fastmcp
httpx
python-dotenv
